// frontend/src/api/model/ReceptionistDoctorScheduleSlotDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LocalDate } from "./LocalDate";
import type { LocalTime } from "./LocalTime";

export interface ReceptionistDoctorScheduleSlotDto {
  publicId?: string;
  date?: LocalDate;
  startTime?: LocalTime;
  endTime?: LocalTime;
  scheduleStatus?: string;
  /** @nullable */
  scheduleStatusColorCode?: string | null;
  doctorPublicId?: string;
  doctorName?: string;
  doctorSpecialization?: string;
  /** @nullable */
  appointmentPublicId?: string | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  visitReason?: string | null;
  /** @nullable */
  appointmentStatus?: string | null;
  /** @nullable */
  appointmentStatusColorCode?: string | null;
}